Don't create MessageBoxes from non-main window
MessageBox use plattform functions and is created on the stack.
While it is shown, it is also part of the return stack.

Trying to close the parent window (MIDI organ load events) while the message box is shown
results therefore in a crash.

Work around by moving the display of the messagebox to a safe location.

Ugly, but the simplest solution.


git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@1371 69509848-2a73-49d5-a893-c4d8c934025a