Introduce USE_INTERNAL_ZITACONVOLVER
This option allows to use a system zitaconvolver library instead of the
bundled one. By default we will still use our bundled internal
zitaconvolver.

Unfortunately as of now, there is no pkgconfig support for zitaconvolver
so for now we have to just assume that zita-convolver is installed in
the system, if USE_INTERNAL_ZITACONVOLVER=Off is used.

Fixes #1095