Use CMAKE_INSTALL_LIBDIR to specify LIBINSTDIR on Linux
The directory for system libraries is a nice conundrum on Linux.
It could be lib64 or lib or lib/<multiarch-tuple>. To sort this out,
CMake provides via GNUInstallDirs the CMAKE_INSTALL_LIBDIR variable,
see https://cmake.org/cmake/help/latest/module/GNUInstallDirs.html
The distributions are then free to set this according to their tastes.