Change GO_USE_JACK usage and link golib to jack
The variable defined via option() are usually not only set to string
values, but to any constant that evaluates to true, for instance like "1",
see https://cmake.org/cmake/help/latest/command/if.html
This makes the usage of "GO_USE_JACK" in this context consistent.

Furthermore, this reverts the change of #1094 and golib now pulls the
dependency on jack. This is more future-proof.